rename MVTTileLayer to MVTLayer
