+++ docs/table-of-contents.json
@@ -100,7 +100,7 @@
             {"entry": "docs/layers/tile-layer"},
             {"entry": "docs/layers/trips-layer"},
             {"entry": "docs/layers/terrain-layer"},
-            {"entry": "docs/layers/mvt-tile-layer"}
+            {"entry": "docs/layers/mvt-layer"}
           ]
         },
         {
