+++ modules/geo-layers/src/mvt-layer/clip-extension.js
