+++ test/modules/geo-layers/mvt-layer.spec.js
@@ -1,14 +1,14 @@
 import test from 'tape-catch';
 import {generateLayerTests, testLayer} from '@deck.gl/test-utils';
-import {MVTTileLayer} from '@deck.gl/geo-layers';
-import ClipExtension from '@deck.gl/geo-layers/mvt-tile-layer/clip-extension';
+import {MVTLayer} from '@deck.gl/geo-layers';
+import ClipExtension from '@deck.gl/geo-layers/mvt-layer/clip-extension';
 import {GeoJsonLayer} from '@deck.gl/layers';
 
 import * as FIXTURES from 'deck.gl-test/data';
 
-test('MVTTileLayer', t => {
+test('MVTLayer', t => {
   const testCases = generateLayerTests({
-    Layer: MVTTileLayer,
+    Layer: MVTLayer,
     assert: t.ok,
     sampleProps: {
       urlTemplates: [
@@ -17,7 +17,7 @@ test('MVTTileLayer', t => {
     },
     onBeforeUpdate: ({testCase}) => t.comment(testCase.title)
   });
-  testLayer({Layer: MVTTileLayer, testCases, onError: t.notOk});
+  testLayer({Layer: MVTLayer, testCases, onError: t.notOk});
   t.end();
 });
