+++ test/render/golden-images/mvt-layer.png
