+++ test/render/test-cases/mvt-layer.js
@@ -1,8 +1,8 @@
-import {MVTTileLayer} from '@deck.gl/geo-layers';
+import {MVTLayer} from '@deck.gl/geo-layers';
 
 export default [
   {
-    name: 'mvt-tile-layer',
+    name: 'mvt-layer',
     viewState: {
       longitude: -100,
       latitude: 40,
@@ -11,13 +11,13 @@ export default [
       bearing: 0
     },
     layers: [
-      new MVTTileLayer({
+      new MVTLayer({
         urlTemplates: ['./test/data/mvt-tiles/{z}/{x}/{y}.mvt'],
         getFillColor: [0, 0, 0, 128],
         getLineColor: [255, 0, 0, 128],
         lineWidthMinPixels: 1
       })
     ],
-    goldenImage: './test/render/golden-images/mvt-tile-layer.png'
+    goldenImage: './test/render/golden-images/mvt-layer.png'
   }
 ];
